[Tests] Move AlphaDIA tests with real data to integration tests
