+++ tests/integration/conftest.py
@@ -1,0 +1,20 @@
+"""Shared logic for integration tests."""
+
+import os
+from typing import Generator
+
+import pandas as pd
+import pytest
+
+from alphabase.tools.data_downloader import DataShareDownloader
+
+
+@pytest.fixture(scope="function")
+def example_alphadia_tsv(tmp_path) -> Generator[pd.DataFrame, None, None]:
+    """Get and parse real alphadia PG report matrix."""
+    URL = "https://datashare.biochem.mpg.de/s/cN1tmElfgKOe1cW"
+
+    download_path = DataShareDownloader(url=URL, output_dir=tmp_path).download()
+    yield download_path
+
+    os.remove(download_path)
